Fixed deprecated numpy aliases

+++ penelope/notebook/co_occurrence/tabular_gui.py
@@ -25,10 +25,10 @@ DISPLAY_COLUMNS = ['time_period', 'w1', 'w2', 'value']
 def empty_data():
     return pd.DataFrame(
         data={
-            'time_period': pd.Series(data=[], dtype=np.int),
+            'time_period': pd.Series(data=[], dtype=np.int32),
             'w1': pd.Series(data=[], dtype=str),
             'w2': pd.Series(data=[], dtype=str),
-            'value': pd.Series(data=[], dtype=np.float),
+            'value': pd.Series(data=[], dtype=np.float32),
         }
     )
